chore: update linter

+++ internal/engine/command/phoenix/faucet.go
@@ -21,7 +21,7 @@ func (pt *Phoenix) faucetHandler(cmd command.Command, _ entity.AppID, _ string, 
 		return cmd.FailedResult("Uh, you used your share of faucets today!")
 	}
 
-	txID, err := pt.wallet.TransferTransaction(toAddr, "Phoenix Testnet Pagu PhoenixFaucet", int64(pt.faucetAmount))
+	txID, err := pt.wallet.TransferTransaction(toAddr, "Phoenix Testnet Pagu PhoenixFaucet", pt.faucetAmount)
 	if err != nil {
 		return cmd.ErrorResult(err)
 	}
